Ordenaçãod e Ppcs por vigente e correção de erros em Curso

+++ app/Http/Controllers/CursoController.php
@@ -161,7 +161,9 @@ class CursoController extends Controller
 
         if ($ato_autorizacao && $ato_autorizacao->isValid()) {
 
-            File::delete(storage_path('app/public/' . $curso->atoAutorizacao->path));
+            if($curso->atoAutorizacao->path) {
+                File::delete(storage_path('app/public/' . $curso->atoAutorizacao->path));
+            }
 
             $diretorio = 'ArquivoAtoAutorizacao';
             $nomeAtoAutorizacao = $ato_autorizacao->store($diretorio);
